+++ Evaluation-Module/src/main/java/Helpers/HelpersMain.java
@@ -8,7 +8,7 @@ import java.io.File;
 public class HelpersMain {
     public static void main(String[] args) {
         FileHelper fH = new FileHelper();
-        int size = fH.getFileFolderSize(new File("/tmp/Evaluation"));
+        long size = fH.getFileFolderSize(new File("/tmp/Evaluation"));
         System.out.println(size);
 
     }
